security_multitool/requirements.txt
typer
requests
cryptography
python-nmap
